[Wave] Split index sequence analysis
This PR splits the index sequence analysis code into
two separate files, one for partition strided operators and the other
for index sequence analysis.
