+++ lit_tests/kernel/wave/index_sequence_analysis.py
@@ -17,10 +17,12 @@ from iree.turbine.kernel.wave.minimize_global_loads import minimize_global_loads
 from iree.turbine.kernel.wave.shared_memory_indexing import (
     apply_shared_memory_indexing_corrections,
 )
-from iree.turbine.kernel.wave.index_sequence_analysis import (
-    partition_strided_operators,
+from iree.turbine.kernel.wave.analysis.index_sequence_analysis import (
     set_node_indices,
     set_post_expansion_indices,
+)
+from iree.turbine.kernel.wave.analysis.partition_strided_operators import (
+    partition_strided_operators,
 )
